+++ app/src/main/AndroidManifest.xml
@@ -174,11 +174,11 @@
             />
 
         <service
-            android:name="com.amaze.filemanager.asynchronous.ftpservice.FTPService"
+            android:name=".asynchronous.services.ftp.FTPService"
             android:enabled="true"
             android:exported="true" />
 
-        <service android:name=".asynchronous.ftpservice.FTPTileService"
+        <service android:name=".asynchronous.services.ftp.FTPTileService"
             android:icon="@drawable/ic_ftp_dark"
             android:label="@string/ftp"
             android:permission="android.permission.BIND_QUICK_SETTINGS_TILE">
@@ -189,7 +189,7 @@
         </service>
 
         <receiver
-            android:name="com.amaze.filemanager.asynchronous.ftpservice.FTPReceiver"
+            android:name=".asynchronous.services.ftp.FTPReceiver"
             android:exported="true">
             <intent-filter>
                 <action android:name="com.amaze.filemanager.services.ftpservice.FTPReceiver.ACTION_START_FTPSERVER" />
